+++ custom_components/life_events/manager.py
@@ -133,6 +133,7 @@ class LifeEventsManager:
             "date_of_death": fields.get("date_of_death", current.date_of_death),
             "icon": fields.get("icon", current.icon),
             "phone_number": fields.get("phone_number", current.phone_number),
+            "time": fields.get("time", current.time),
             "attributes": fields.get("attributes", current.attributes),
         }
         updated = Event.create(**merged)